Really make refcount-opts deterministic
Summary:
Key contains a Block*, so sorting on it isn't entirely
deterministic. I've seen a few cases where the order did change,
although its rare (presumably because blocks are allocated from
Arenas), and it only affects ir-dumps as far as I can tell (we still
put the same incs and decs, in the same order in each block - the
difference just shows up in which blocks we process first, resulting
in different instruction numbering in the ir dumps).

Reviewed By: ricklavoie

Differential Revision: D18416600

fbshipit-source-id: 388840e83dc4a708f6bf496e004dacdb814e2778