adding some missing expand_type
Summary:
Whenever we use `simplify_unions`, we would assume the result would be an already expanded type. This is not always the case anymore, so we need to add some calls to `expand_type`, especially in the `expand_type_and_xxx` functions.

At the end of this stack, D18059114 will deal with removing the unnecessary calls to `simplify_unions` (the only legitimate ones are the ones defining the `~on_tyvar` argument), so in most cases, only an `expand_type` will remain.

Reviewed By: andrewjkennedy

Differential Revision: D18061405

fbshipit-source-id: ec37db30f6e810f3d0d6ea0b1c3de9e003e37db2