introduce env.tyvar_occurrences
Summary:
Unions and intersections containing unsolved type variables may remain in an unsimplified form once those type variables get solved.

For example, consider the union `(#1 | int)` where #1 is an unsolved type variable.
If #1 gets solved to int, then this union will remain in the unsimplified form
`(int | int)` which compromise the robustness of some of our logic and might cause performance issues (by creating big unsimplified unions).

To solve this problem, what this series of diffs does is wrap each union and intersection in a type var, so we'd get here `#2 -> (#1 | int)`, and register that #1 occurs in #2. Then when #1 gets solved, we simplify #2.

This first diff introduces the map `tyvar_occurrences` in the environment which allows to register which var occurs in which other var.

Reviewed By: andrewjkennedy

Differential Revision: D18061338

fbshipit-source-id: 055ae5a8c26e1ce89375849025ad4e1152cf2760