Always create the IRUnit's AnnotationData
Summary:
There was just one place trying to avoid creating it in some cases,
and the logic was incorrect as it didn't cover all annotations that
may be inserted in the IRUnit. It doesn't cost much to create an
AnnotationData, so just do it. This is in the func-body dispatch,
in which I don't think we currently annotate anything, so it wasn't
causing any problems.

Reviewed By: mofarrell

Differential Revision: D18412552

fbshipit-source-id: 188426fab7e84632083b24d31bc26e5e24b2da78