Make it possible to tell what can be foreached with one `is` expr.
Summary:
There is currently no sane way to test from hack whether something can be iterated with forech. This is because we have both \Traversable and \HH\Traversable, as well as both \Iterator and \HH\Iterator. The latter pair is connected via inheritance (\HH\Iterator extends \Iterator) but the former pair is not. Some things that can be iterated are \HH\Traversable but not \Traversable (e.g. all arrays), some things that can be iterated are \Traversable but not \HH\Traversable (e.g. things like DOMNodeList that implement \IteratorAggregate and don't get \HH\Traversable by some other mechanism like using StrictKeyedIterable; and things that directly implement \Iterable (only ResourceBundle)). Finally some things are _both_ \Traversable and \HH\Traversable (e.g. all hack collections, generators, ArrayObject).

IMO the best first step towards fixing this is to:
1) Make \HH\Traversable extend \Traversable (as \HH\Iterator extends \Iterator)
2) Stop implementing \Traversable and \Iterator directly (only via the HH versions).

This way things that had \Traversable or \Iterator typehints continue to work as before, but we can now do any of the tests `$foo is \Traversable`, `$foo is \HH\Traversable`, or `$foo is Traversable` (the last is the same as the second due to auto-importing). All of these will now return true iff $foo can be iterated with foreach.

Reviewed By: kmeht

Differential Revision: D18425435

fbshipit-source-id: 30e50a5f18c11044fb0532fad9b9c3850fdf5827