Seal \Traversable and \Iterator.
Summary:
Step 2: seal \Traversable and \Iterator so that we can't accidentally backslide and start extending them directly.

This is a breaking change for OSS. People that were extending/implementing these need to either remove the leading \ or add \HH in order to get the derived class (via auto-import or explicitly).

Reviewed By: kmeht

Differential Revision: D18425490

fbshipit-source-id: 4798231a59a3eed342c44dc2c5d355fbd28cadd6