Fix bugs from the IterArgs refactoring
Summary:
This change fixes two bugs in D18362188: one that prevented Eval.DumpHhas from pretty-printing optimized bytecode, and one that prevented DCE for key-value IterInit -> value-only IterInit.

For the HHAS bug, valid HHAS doesn't include any IterArgs::Flags, because these flags are only set by HHBBC. However, we also use the disassembler to implement Eval.DumpHhas for optimized bytecodes, and iterators may have flags for this usage. We make the new HHAS printing (for IterArgs with flags set) a strict superset of the old one: if there are no flags, we don't print the flags list, but if at least one is set we will. This way, if we try to load this extended HHAS, we'll crash. There are already flags in FCallArgs that can't be round-tripped through HHAS, so this behavior isn't new.

For the DCE bug, I incorrectly examined the iterator state before IterInit to determine whether to DCE - this state is always going to be invalid.

Reviewed By: binliu19, mxw

Differential Revision: D18441933

fbshipit-source-id: 5287c293863303f8533c582ce045eea444e2a765