Fix aliasing bug when global_size is 0
Summary:
The shmem code to read, write, and clear global storage dereference
`global_storage[0]`, which holds the size of the serialized data.

However, if the `global_size` configuration is set to 0, then
`global_storage[0]` aliases `deptbl[0]`. Once these things are aliased, we can
easily introduce confusion.

This diff ensures that `global_storage[0]` is always allocated, even if
`global_size` is 0.

Reviewed By: mroch

Differential Revision: D18382420

fbshipit-source-id: e94df791df4201c207ad1826b74adca13cd62338