One more case for iter docs
Summary: I missed the collections case when updating these docs. Now that we have this case, it's easier to describe the "next" behavior in terms of the type of the iterator instead of the type of the base. There are two cases - collections -> array iterator and IteratorAggregate -> object iterator - where we coalesce "next" behavior by creating the same kind of iterator for different kinds of bases, so naming the types of iterator is a clean way to reduce this casework.

Reviewed By: mofarrell

Differential Revision: D18457443

fbshipit-source-id: ceeb8187718289f4c98fcc7ae137e648cd52a6eb