Fix parsing of aliased_namespaces JSON as object (not array)
Summary:
The JSON parser in Hhbc_options expects hhvm.aliased_namespaces to be:
- JSON array if empty
- JSON object in general (non-empty doesn't parse as array!)

Modify OCaml code to no longer accept empty array as it's redundant & confusing.

In Rust, expect it to always be JSON Object, since we cannot support
both without exorbitant logic (also not sure if we should, as the array
was probably a leftover from some old migration).

Reviewed By: shiqicao

Differential Revision: D18438723

fbshipit-source-id: d9d62ac2f1030cbadc5da093efa8a5f292f364a0