Disable AST constant folding for numbers with trailing garbage
Summary:
Out of two callers of `string_to_int_opt`, one has this option already hardcoded to false, and the other one matters only for performance (we end up not folding some constants).

Since allowing garbage at the end of numbers is a PHPism and should not actually happen in any places where performance matters, we can just remove it to simplify the code.

Reviewed By: losvald

Differential Revision: D18411005

fbshipit-source-id: ec9cc0f7bdf340239f7e4a7451631ad39bb3f95b