add a naming table to simple-repo
Summary:
I'll be wanting to write tests which use a naming table for simple-repo.

This diff adds a buck target which generates the naming-table for that repo. That way, my future tests can use `$(location ...)` and use the naming table. And if lots of tests all want to use the naming-table, well, buck will share the it for all of them.

Reviewed By: jewelpit

Differential Revision: D18408420

fbshipit-source-id: b071d6980257d3937feb794cab348e5221b75b7f