make output deterministic
Summary:
I'll want to write integration tests for hh_decl.

The pattern we've used elsewhere in verify.py is that some things with are non-deterministic (e.g. the output of Hh_logger.log with dates embedded) get emitted to stderr, and we tell verify.py to ignore stderr.

Names.rs currently prints "loading name index from <absolute-path>", and absolute paths paths are non-determinstic (depending upon who is executing the code or where). So I've moved that to stderr.

Typecheck.ml currently prints the TAST, but this has a load of absolute paths in it. I've changed it to only print TAST to stderr.

Typecheck.ml currently prints errors with absolute paths. I've changed it to only print basenames for errors.

Together, all these changes are enough to make hh_decl stdout be deterministic.

Reviewed By: jewelpit

Differential Revision: D18416916

fbshipit-source-id: de5578297f8be50f1c289e6e04fe34d6836ab101