inproc
Summary:
I'm moving towards being able to write integration tests. From past experience, I don't want those tests to fork any processes - I want them all to be inproc. That means (1) workers have to be inproc with the orchestrator, (2) decl-service has to be inproc with the workers.

This diff does the ocaml side of things. It adds a new command-line option to the orchestrator:
```
hh_mapreduce typecheck --root ... --inproc <namingtable>
```
(The naming table is needed here because, since we're doing decls inproc, we need to know where to fetch them from).

Because we're doing (1) workers inproc with the orchestrator, I've had to factor out the core code of the worker into `typecheck.ml:run_worker_core`. This will be called directly from the orchestrator (in case of inproc) or called from the prototype (in case of normal behavior).

Reviewed By: jewelpit

Differential Revision: D18373041

fbshipit-source-id: f2efd297e5efbb942aefb44755cd2fc5337d9722