optional mode-flag arg
Summary:
Our verify.py script invokes its specified program with the filename of the test case as its first argument, and all other flags as the rest.

This is great for tools like `hh_single_type_check`, but it doesn't work for commands like `hh_client check` or `hh_client start` which take a "mode" as their first flag.

This diff changes verify.py to take an optional `--mode-flag` paramater, to specify the mode flag that will be passed first to its specified program.

Reviewed By: jewelpit

Differential Revision: D18416881

fbshipit-source-id: b84bd86054116d9a211f1e2538b7abbc240f5cef