Fix and cleanup handling of get{pw,grp}{uid,nam}_r*()
Summary:
These glibc calls return non zero on error. Otherwise, they set the result
pointer to the provided buffer or nullptr depending on whether the entry was
found.

Most of the call sites in HHVM were reading the result from the provided buffer
instead of via the result pointer. Some of them forgot to check whether the
result pointer was nullptr, causing to potentially read garbage instead of
segfaulting on null dereference.

Make this safer by creating a structure that encapsulates all the buffers
passed to these calls, making it easier to consume data via the result pointer.

Access the result via the result pointer and make sure it is non nullptr.

Reviewed By: ricklavoie, alexeyt

Differential Revision: D18442830

fbshipit-source-id: 0a9096a9413cbd9cef3f6b6f23b28380bcd0a359