rework Typing_expr_dep_type.make
Summary: Looking at the history of this function, first `should_apply` was created, which dealt with unions, then later on `apply` started dealing with unions too. Nowadays, it seems the two functions can be unified. Currently, with certain unions or intersections or combinations of both, we create expression dependent types for non-class types like `nonnull`, which doesn't make much sense to me. This diff unifies both `should_apply` and `apply` in such a way that we "apply" directly only on class types which need attention (non-exact, non-final, ...).

Reviewed By: dlreeves, andrewjkennedy

Differential Revision: D18229942

fbshipit-source-id: 91933940c4ad9d6f4dc7ba64de89663d1b765ad6