Remove some float parsing / printing from TypedValue / Ast_constant_folder
Summary:
There is absolutely nothing suggesting that OCaml's scanf / printf functions have anything in common with HHVM's way of treating floats, and they are rather obviously inconsistent, leading to differences in behavior of the same code depending on whether constants end up folded or not. The bugs / effort to eliminate them is not worth the marginal benefit it can provide.

There are more calls to Hhbc_string_utils.Float / float_of_string that likely will also be problematic and it will be easier to remove them from emitter than try to fight it, but leaving them out for now until we have time to asses their impact too.

Reviewed By: shiqicao

Differential Revision: D18457936

fbshipit-source-id: 088b86603866255d71a8adf1b713d7939a1c8323