Update DefInlineFP's marker when it is pushed to a catch trace
Summary: On D18332484, I noticed that `vmpc()` was not being updated correctly in catch traces so I used the `DefInlineFP`'s marker to update `vmpc()`. However, as it turns out, `DefInlineFP` marker, even when it is pushed to a catch trace by pdce inlining, it still points to the beginning of the function. This diff updates `DefInlineFP`'s marker to `BeginCatch`'s marker so that we correctly update `vmpc()`.

Reviewed By: paulbiss, jano

Differential Revision: D18431993

fbshipit-source-id: 6387acd32497aec4c49cebdb4e0f1b3ae5df7c7c