Change emit_symbol_refs to expect Hhbc_id.t's (with bug fix)
Summary:
Expect that names for symbols that need to be emitted in emit_symbol_refs need to first be passed through Hhbc_id.from_ast_name, which modifies the string to a format that HackC/HHVM expect.

Previous diff failed due to shadowing behavior that wasn't caught - this diff hopefully catches all instances.

Reviewed By: ricklavoie, kmeht

Differential Revision: D18465326

fbshipit-source-id: 7f766858a4b15aaacda3d34bf78ce79c93fb187e