Add support for async and generator functions.
Summary:
This diff adds support for async functions, which are signified by the async keyword, and generator functions, which are signified by having a yield anywhere inside their body.  Async is easy, but handling generators is a little more complex.  When we see the function keyword, we push a `false` value onto our yield stack.  When we see the yield keyword, we convert the topmost value to `true`.  Then, when we create the function, we pop the topmost value and use that to determine whether we saw a yield.

We could use a single value instead of a stack, but this will make it easier to avoid false positives when a non-generator function contains a generator lambda.

Differential Revision: D18361712

fbshipit-source-id: d38ae94175c87bb49c30865769dca0cf2e0c079d