Remove another redundant add_ns
Summary:
As part of the work on making namespace elaboration consistent, we discovered that the nast_check visitor was visiting a tree where function bodies had not yet been named and therefore names were not elaborated.

To work around this, we simply added some workarounds that probably didn't play nicely with true namespacing behavior, such as appending the global \ in front of names to perform special name checks.

This diff removes those unnecessary appends now that namespace elaboration makes sense.

Reviewed By: Wilfred

Differential Revision: D18454377

fbshipit-source-id: 4450dcfa60bddecadbb2e619a4f8e541407faf54