Use IterArgs in the IterData IR data struct
Summary:
The IterData / IterInitData IR extra data structs just carry the same information as the new IterArgs struct. Modify the code to reflect that, and update bytecode.specification to explain IterArgs.

We also drop the IterTypeOp HHBC subop, but for now, we keep this enum in array-iterator.h as a way to burn in the appropriate "new_iter_array" call into the JIT. We'll drop this enum soon, but when we do so, it will have performance effects that I don't want to consider as part of this change.

Reviewed By: mxw

Differential Revision: D18437640

fbshipit-source-id: 43f7d0c82d9330fe60ae5d415378e66a82ccdde0