Make context version of find refs
Summary:
Based on the performance improvements we've seen so far, I think there's even more fast to be had.

We found that DocumentHighlight was one of the functions that wasn't speeding up like we hoped.  That's because it uses FindRefsService, which didn't have a context method of its own.

Here's what this diff does:
* Break out FindRefsService.find_refs into an inner function and a shell
* Implement a new method in IdentifySymbolService that reuses memoized symbols out of a TAST
* Add a new FindRefsService.find_refs_ctx method that uses provider context and the inner function, plus the memoized symbols
* The context-based approach requires both a FileInput object and a filename object, so let's modify everything that needs it
* We can eliminate the old non-context version of ServerHighlightRefs.

Reviewed By: arxanas

Differential Revision: D18466649

fbshipit-source-id: a8189bd83b1e3bd7b485c9db1093bc11b5d21ceb