Pass `Namespace_env` as explicit parameter into `typing_get_locals`
Summary:
The namespace environment parameter is part of the accumulator tuple. However, it's not actually accumulated; it's strictly used as an input parameter. This commit changes it so that `nsenv` is an explicit parameter as part of the tuple.

It simplifies a lot of the code that just discards the `nsenv`. It also makes it easier for me to thread through a `Provider_context.t` as an additional explicit parameter in a later diff.

Reviewed By: kmeht

Differential Revision: D18515520

fbshipit-source-id: d1d450a9679b4473573b68ae41d4603c41a60164