simplify unions after solving type variables
Summary:
In the smart constructor for unions Typing_union.make_union, if the result is a union which contains unsolved type variables, then we wrap it in a type variable.
Do something similar with `make_intersection`

When a type variable gets solved, unions containing this type variable may be solved in `Typing_type_simplifier` if all type variables from this union are solved.

This is because there are a lot of types like
```
( Vector<#1> |
  Vector<#2> |
  ...
  Vector<#n> )
```
If each time a type variable gets solved, we attempt a union simplification which is quadratic in n, we get bad perf.

Therefore, we only simplify a type when all type variables in it are solved.
To that end, we also maintain a map `tyvars_in_type` in the environment that is the dual of the other map `tyvar_occurrences` introduced in a previous diff. Furthermore, we now maintain the invariant that we only keep the occurrences links if the type variable that occurs is unsolved or itself contains unsolved type variables.

So for example, for
```
#1 -> (#2 | #3)
#2 -> (int | #4)
#4 -> string
#3 -> #5
#5 unsolved
```
We'd have recorded that #5 occurs in #3 and #3 in #1, but since #2 and #4 are solved and do not contain any unsolved type variables, we don't record their occurrences.

We can now simplify a type variable type only when their entry in `tyvars_in_type` is empty.

With the previous example, when #5 gets solved, if it gets solved to something that still contains unsolved type variable, we don't do any simplification, but if it gets solved to say int, to maintain the invariant described above, we'd remove it from the types in #3, and since #3 now wouldn't contain any more unsolved type variables we could simplify it, then recursively remove it from the types in #1 and finally simplify #1 itself.

We also remove now unnecessary usages of simplify_unions

Reviewed By: andrewjkennedy

Differential Revision: D18059114

fbshipit-source-id: 73eb8c40087d519246306382b9c2ab99ad1deb20