Throw exceptions when updating fails
Summary:
When the autoload map fails to update today, we get a cryptic error message:

```
Failed to update native autoloader, not natively autoloading <path_to_www_repo>.
```

We don't get any clues as to *why* we failed. We can trace this lack of information back to my decision to either return a `bool` or throw an exception when updating fails, depending on the specifics of the failure. This was a pretty terrible design decision, so I'm fixing it here. Now we always throw an exception on failure, which means we always have information to log.

Reviewed By: dneiter

Differential Revision: D18495884

fbshipit-source-id: 16d6ad0b4740ad19c39acbd2c24d03db07640fad