Fix bad error position reported when type checking new object expressions
Summary:
Consider the following example (added as a new test case below):
```
//// file1.php
<?hh
function test<T as C>(
  T $c,
  A $a,
): void {
  $b = $c->b();
  $_ = new $b($a);
}
//// file2.php
<?hh
interface A {}
final class B<Ta as A> {
  public function __construct(Ta $a) {}
}
interface C {
  abstract const type Ta as A;
  public function b(): classname<B<this::Ta>>;
}
```
At the moment, type checking of the expression `new $b($a)` in file1.php produces an error with the primary position in file2.php (pointing to `this::Ta` in the type signature of the method `b`).  This diff fixes this issue.

Reviewed By: Matt-Schellhas

Differential Revision: D18480500

fbshipit-source-id: 7c7bb25631719d501cae3e1798ae08d09bdd889f