Get closer to true out-of-source builds for Hack
Summary:
It should be possible to build HHVM and Hack from an entirely different directory
to the source, and for the source directory to be completely unchanged.

This is expected by Debian build scripts, for example.

This gets us /almost/ all the way there - but Dune still generates `.merlin`
files in the source directory. https://github.com/ocaml/dune/issues/2568 to address that.

Reviewed By: jjergus

Differential Revision: D16944399

fbshipit-source-id: b0b4fc17870dfda95d3d555928a0c99911fdeee0