TypedValue
Summary: A subset of `typed_value.ml` sufficient to make implementation of `Ast_constant_folder.fold_program` produce identical results on all test files in `hphp/` (module float printing issues from D18439150). This doesn't have everything, and we'll need to come back to it when we want to use it in other callsites, but I don't want to do it prematurely without having code to test it on / deciding whether it may be easier to rip some of it from OCaml side as in previous diffs of this stack.

Reviewed By: losvald

Differential Revision: D18469000

fbshipit-source-id: 49cc30d0408e6df9bc2f9436100347eee3525f67