Use BitsetArray in refcount-opts
Summary:
Reduces the size of each of the PreBlockInfo bitvectors from 32 to 16,
and does a single allocation for all the bits (rather than one per
bitvector per block).

We could take this a step further, and just store the base BitsetRef
for each block, and generate the others on the fly, but for now this
is a big improvement with little impact on the uses of PreBlockInfo.

Reviewed By: ricklavoie

Differential Revision: D18496257

fbshipit-source-id: ef8b355d5b3135fb444eb93beb68c14a5f7b832e