Added unit test for current behavior
Summary:
Actual feature to come in next diff. This is a test for current behaviour which'll be updated too.

Three basic tests that check that hovering over a literal (string or int) in or outside of a function invocation returns 'string' or 'int' as appropriate. I also checked the current behaviour manually by opening the file 'hover.php' and checking that is what is shown.

Caveats:
- I wrote a pad_left() function since Str\ namespace isn't being recognized. This may be a fbcode vs www issue?
- The 'range' key inside the result dict is apparently not produced. This seems inconsistent with a constant to me.

Differential Revision: D18079314

fbshipit-source-id: b840c986983227e53ec4fff34c58524d73a59ac3