Fix invariant() argument checking
Summary:
We desugar calls to invariant() into the condition check, then (if
failing), a call to invariant_violation(). If someone calls
invariant() with just the condition (no format string), then we allow
it if the condition evaluates to true. If it evaluates to false, we
emit an error about supplying 0 arguments to invariant_violation(),
which is confusing.

Make this more consistent by requiring the format string in
invariant(). Moreover, refrain from doing the desugaring until there's
at least two arguments supplied.

The type-checker already enforces that invariant() is always called
with a format string, so this brings the runtime closer to the
type-checker.

Reviewed By: paulbiss

Differential Revision: D18575929

fbshipit-source-id: f9ec9e167a1ec2758513f14bbad5080c535c68ff