Move iterator output locals
Summary:
Suppose we have a function that calls another function on every element of an iterable object $base, such as the Hack standard library method "Vec\map". For such a function, we currently emit bytecode like the following (catch traces elided for clarity):

    NewVec
    PopL $result
    CGetL $base
    IterInit $value L1
    L0:
      CGetL $value
      FCall "InnerFunction"
      BaseL $result
      SetM 0 W
      IterNext $value L0
    L1:
    PushL $result
    RetC

This bytecode is inefficient because it does an extra inc-ref and dec-ref on $value in the loop. The first inc-ref is in the IterInit / IterNext, where the output local is set, and the second is in CGetL, where it's prepped as an argument for InnerFunction. We can optimize this code by converting the CGetL into a PushL; this eliminates one inc-ref (at that bytecode) and one dec-ref (at IterNext / RetC, depending on if there are more elements or not).

In this change, we modify HHBBC's DCE to allow this optimization. Specifically, we note that iterator ops never really "read" their output locals - at least, not in a sense that matters for DCE analysis. A local is live going into an IterInit / IterNext iff it's live in the loop done block (here, L1), or the loop catch block; if we end up taking the loop body block (here, L0), we're going to overwrite this local without reading it. Hence, we can treat $value as being dead prior to both iterator ops in our example. This optimization is nice because it doesn't even rely on the base being an array; the same logic applies to subclasses of Iterator. (I previously tried modifying the semantics of IterInit / IterNext to not inc/dec their outputs for array-like bases, but it took a lot more code, was quite questionable semantically, and I couldn't get frame cleanup to be correct after suppressing refcounts.)

We implement this optimization by modifying the how we push the in-state of the loop body block L0 to the out-state of the blocks ending in IterInit / IterNext. Assuming that this block is not also the loop end block L1, we can kill the locals that the iterator op will set. The in-state of the exception block is tracked separately; we don't kill locals for it. Because we no longer track these variable's liveness within the loop body, we can't do iterator key DCE anymore, but this new change is more useful for typical code. (Iterator key DCE is also handled by JIT-ing specialized iterators, and this change will let us do more of that.) We could bring back iterator key DCE by tracking the next and taken block's in-states separately, but that complexity belongs in a follow-up.

Reviewed By: mofarrell

Differential Revision: D18572666

fbshipit-source-id: 5b9218b62e60a92a687f337fc78f5d4068fa5d0f