Skip funcPrologueRedispatch when it's statically proven to be unnecessary
Summary:
If we don't know the target of a call, we instead call the
funcPrologueRedispatch unique stub, which will take care of shifting
the passed arguments around to match what the callee expects. However,
this is only necessary if we don't know if the number of passed
arguments exceeeds the number of non-variadic declared arguments of
the callee.

Even if we don't know the target of a call, static analysis in HHBBC
may still be able to provide a lower-bound on the possible number of
non-variadic arguments of the potential callees. If the number of
passed arguments is less than this bound, we can skip the
funcPrologueRedispatch stub entirely, and call the target's prologue.

Add an extra boolean to the FCallArgs to indicate whether HHBBC has
proven that the stub can be skipped, and add logic to HHBBC to set
this flag as appropriate. Since the common case is that we can skip
the stub, encode the boolean so that we only pass a space penalty if
it's not set. Use the flag in irlower-call to skip calling the stub,
and instead load the proper prologue pointer out of the Func and call
it directly. This necessitated inventing a "callphpr" instruction
since we've never done a PHP-abi call with an indirect register
before.

Reviewed By: ottoni

Differential Revision: D18582662

fbshipit-source-id: 5440ddc6681a35280f103c892d711b580f124f85