avoid recursion depth counter
Summary:
In Ocaml lowerer `pDef` tries to lower nodes as statements if nothing is matched, similarly `pStmt` tries to  lower nodes as inline def if nothing is matched. This will cause stack overflow. `env.max_depth` is used to avoid overflow and somehow it was set to 42.

Bouncing between `pDef` and `pStmt` 42 times isn't the right way to lower nodes. One of them can pass a location context to another and terminates a recursion.

Reviewed By: dabek

Differential Revision: D18507443

fbshipit-source-id: 2689e9a716e315b7b6698243826a42b68cf7cea2