Fix name mangling for opaque type aliases
Summary: As it turns out, thee name mangling schema for type structures did not take into account the alias and the opaqueness, i.e. `int` and `FBID` both mangled to the same name which resulted in two type structures sharing the same key on the type structure map so the second one never got written.

Reviewed By: jano

Differential Revision: D18597358

fbshipit-source-id: c5f1f84ca1d0448c51126e7aa74920fcf2659a2d