Fix a bug in refcount-opts
Summary:
D14119219 changed the behavior of irrelevant_effects to ignore
GeneralEffects unless it could touch AHeapAny, on the grounds that
refcounts are stored in the heap.

The problem with that is that loads and stores described by
GeneralEffects are assumed to have php semantics; ie loads incref
their result, while stores decref the old value and incref the new -
so even if they only affect the stack, or locals, they still affect
refcounts.

In addition, the diff stopped considering loads at all.

Apparently, the main goal was to make FinishMemberOp be "irrelevant",
and that *can* safely be achieved by ignoring kills (since kills never
affect refcounts). While we're here, moves is required to be a subset
of loads, so we can safely ignore that too.

Reviewed By: paulbiss

Differential Revision: D18596811

fbshipit-source-id: 4af0da6aa068c07761641bf440656490dcfebed5