Make context version of autocomplete
Summary:
As part of the ongoing work to make everything use the context/entry pattern, let's make a context-based version of autocomplete.

As a side note: This standardization helps to eliminate a lot of duplicate versions of this code path.  We were doing file extraction / context generation in lots of different places.

We can also replace the old "char_at_pos" logic and detect that from our autocomplete context task, which already examines this sort of thing.

Reviewed By: arxanas

Differential Revision: D18583408

fbshipit-source-id: 9ad92053a8266f6ab66d3e749b655c08d3b509cd