Fix parser benchmark tool.
Summary: Somehow we were either already pulling in the OCaml stubs and then we weren't, or we didn't need them and then we did, or *something,* but either way this stopped building after D18207462.  This change uses the Rust `cc` package to build the stubs in order to get this project to build again.

Differential Revision: D18533291

fbshipit-source-id: ab6d00e60b0795ba8880498ef4f3400d030acad2