Add proper iteration methods to Node_ in direct decl parser.
Summary: Sometimes we want to iterate over a node and pull out its inner members, but sometimes we just want to use it in a read-only way.  Adding a proper implementation of `iter` not only lets us avoid the allocation of the intermediary vector for non-list cases, it also lets us iterate over a node multiple times.

Differential Revision: D18512209

fbshipit-source-id: 3beb4f74feeb8c8b463319fb0084c85b2058a864