Refactor AST for Call and New nodes
Summary:
By parse error, only a single argument to a function call is allowed to use the splat operator, and it must be the last argument. However, in the AST, this argument is represented as a list, same as the standard arguments. This list is only constructed with a single element and things that match the list only pull out a single element. This diff changes said list to be an option. The affected nodes are `Call` and `New`.

Predictably, this change touched nearly every corner of the Hack codebase. Given the choice of `Option.value_map` and an explicit `match` expression, I've chosen what I consider to be the more readable option on a case-by-case basis. In some cases I've introduced auxillary functions to preserve the style of the code in question. Note: Zoncolan has their own implementation of Option where `value_map` is named `map_default`.

Reviewed By: manzyuk

Differential Revision: D18604077

fbshipit-source-id: 000bc30b7a732b89236bdfb8ce7ee66d531269fb