Remove uses of try_add_err
Summary:
The `Typing_ops.sub_type` function calls `Typing_subtype.sub_type` and wraps any resulting errors with a top-level "principal" error determined by the position and `ureason` parameters. This is stateful and brittle - for example, it behaves badly if we defer error handling until a later solving of constraints.

Better to construct the full error in the first place. We do this by wrapping the `on_error` callback that is passed into `Typing_subtype.sub_type`. Although this is still a "wrapping", it is not stateful.

A slight complication is that many of the errors generated inside subtype checking override the error code that was determined by the `on_error` parameter e.g. reactivity mismatch. To deal with this, the `on_error` callback now takes an optional error code parameter.

Reviewed By: CatherineGasnier

Differential Revision: D18083405

fbshipit-source-id: af8842d22b13eb4bb0cc22199c45b6b381afbbd3