Remove some uses of Errors.try (3)
Summary:
As a first step towards eliminating the global error list, we are removing the use of Errors.try and friends.

Here we use the `on_error` callback instead of `Errors.try`. Unfortunately we are no longer able to deduplicate errors; we need to investigate why duplicates are produced in the first place.

Reviewed By: manzyuk

Differential Revision: D18551547

fbshipit-source-id: 1676c1d6f2fa718aeb5c7c85c2275f67177217f6