Log any exceptions that escape from instruction irlower functions.
Summary:
We sometimes std::terminate in HPHP::jit::irlower::lowerUnit, which is marked noexcept. The line number is variously reported as 117 (the first line of lowerUnit) or 65 (the IR_OPCODES that is building up the switch body in cgInst). It appears that everything up to cgInst has been inlined into lowerUnit, and one of the cg helpers is allowing an exception to escape.

Wrap the call in a try/catch so that we can turn the std::terminate into an assertion failure and log more information about the exception that is causing us to crash.

Reviewed By: oulgen

Differential Revision: D18634411

fbshipit-source-id: be525718ce3ba4d5deb86d85c05e82894583af69