introduce simple unions and intersections for constraint types
Summary:
Before this diff, to do the union of null and Thas_member, we create a new type variable #1 with upper bound Thas_member and return ?#1. This gives rise to multiple tricky stack overflow patterns due to interference with the case in `simplify_subtype` transfoming (nonnull & t1) <: t2 into t1 <: ?t2.

To remediate this, we introduce a TCunion and TCintersection to allow to build simple unions and intersections with constraint types. These are not full-blown unions (you can't union 2 constraint types for example), so their implementation is relatively simple.

NB: for now, such unions and intersection only ever involve null or nonnull in the local type part. If this changes, I can't guarantee that the code for union_i and intersect_i constructing those will not blow up.

Differential Revision: D18573887

fbshipit-source-id: 29d9adce960f07a1e9412f6366c24c79a82d74fc