better detection of early solve failure
Summary: In expand_type_and_solve, we detect whether we solve invariant type variables to nothing. The logic however is brittle. For example, if the type at hand is (#1 & nonnull) and #1 gets solved to nothing, it would fail to detect it, and therefore fail to fail. Same goes if the type at hand is (#1 | #2).

Reviewed By: andrewjkennedy

Differential Revision: D18635991

fbshipit-source-id: 33ce9d1b0630b1a8bf01ecdccf8d05363ce16072